set up bulma, nav and footer
