+++ client/package.json
@@ -10,6 +10,7 @@
     "@testing-library/user-event": "^7.1.2",
     "apollo-server-express": "^2.19.1",
     "bootstrap": "^4.4.1",
+    "bulma": "^0.9.2",
     "jwt-decode": "^2.2.0",
     "react": "^16.13.1",
     "react-bootstrap": "^1.0.1",
