+++ client/public/index.html
@@ -22,11 +22,14 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-  <link href="https://fonts.googleapis.com/css2?family=Rubik+Mono+One&family=Rubik:ital,wght@0,400;1,300&display=swap"
-    rel="stylesheet">
-
-
-  <title>Title</title>
+  
+    <!-- fonts -->
+    <link rel="preconnect" href="https://fonts.gstatic.com">
+    <link href="https://fonts.googleapis.com/css2?family=Rubik+Mono+One&family=Rubik:ital,wght@0,400;1,300&display=swap"
+      rel="stylesheet">
+    <link href="https://fonts.googleapis.com/css2?family=Martel&family=Roboto:wght@400;700&family=Yeseva+One&display=swap"
+      rel="stylesheet">
+  <title>Book End</title>
 </head>
 
 <body>
