+++ client/src/App.js
@@ -2,9 +2,11 @@ import React from 'react';
 import { ApolloProvider } from '@apollo/react-hooks';
 import ApolloClient from 'apollo-boost';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
-import Navbar from './components/Navbar';
+import Navbar from './components/NavBar';
+import Footer from './components/Footer'
 
 // IMPORT PAGES HERE 
+import Home from "./pages/Home";
 
 const client = new ApolloClient({
   request: operation => {
@@ -24,10 +26,11 @@ function App() {
         <>
           <Navbar />
           <Switch>
-            <Route exact path='/' />
+            <Route exact path='/' component={Home} />
             <Route exact path='/saved'  />
             <Route render={() => <h1 className='display-2'>Wrong page!</h1>} />
           </Switch>
+          <Footer />
         </>
       </Router>
     </ApolloProvider>
