+++ client/src/components/Footer/index.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Footer = () => {
+    return (
+        <footer className='footer'>
+            About Us
+        </footer>
+    )
+};
+
+export default Footer;
