+++ client/src/index.js
@@ -2,6 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './index.css';
+import './bulma.min.css';
 import App from './App';
 
 ReactDOM.render(
