+++ client/src/pages/Home.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import Hero from '../components/Hero/index';
+import Genres from '../components/Genres';
+
+const Home = () => {
+    return (
+        <>
+        <Hero />
+        <Genres />
+        </>
+    );
+};
+
+export default Home;
